% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{calc_ci}
\alias{calc_ci}
\title{Create CI for plotting}
\usage{
calc_ci(mean_i, sd_i)
}
\arguments{
\item{mean_i}{The mean}

\item{sd_i}{The standard deviation, already scaled for sample size}
}
\value{
A vector containing the 95\% CI
}
\description{
Create CI for plotting
}
\examples{
calc_ci(0, 1)
}
