% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{clustersum}
\alias{clustersum}
\title{Summing Data based on Clustering Data set}
\usage{
clustersum(training, testing, clustdat, clustcol)
}
\arguments{
\item{training}{Irish_Training_Data}

\item{testing}{Irish_Testing_Data}

\item{clustdat}{Clustering Data (with ID column in first position)}

\item{clustcol}{Column of Clusters}
}
\value{
Training and Testing Data sets with Total Energy Consumption for Each Cluster, and split Extra Info into Training and Testing
}
\description{
Summing Data based on Clustering Data set
}
