% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{dataprocess}
\alias{dataprocess}
\title{Predictor and Model Matrix creator}
\usage{
dataprocess(sumobj, cluster, time)
}
\arguments{
\item{sumobj}{Output of clustersum function}

\item{cluster}{Cluster to Model over}

\item{time}{Time point to Model over}
}
\value{
Predictor, Model Matrix, and Column Means. Class 'dpobj'.
}
\description{
Predictor and Model Matrix creator
}
