% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-dopar.R
\name{\%dopar\%}
\alias{\%dopar\%}
\title{dopar operator}
\usage{
obj \%dopar\% ex
}
\arguments{
\item{obj}{foreach object used to control the evaluation of ex.}

\item{ex}{the R expression to evaluate.}
}
\description{
See \code{foreach::\link[foreach:foreach]{\%dopar\%}} for details.
}
\keyword{internal}
