% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{estimate}
\alias{estimate}
\title{Mean Estimated Data}
\usage{
estimate(sumobj, modelsclusttype, cluster)
}
\arguments{
\item{sumobj}{'sumobj' from clustersum()}

\item{modelsclusttype}{list of models from certain clustering method}

\item{cluster}{cluster to estimate}
}
\value{
Column Means, Estimate (centered), True Value. Class 'estobj'
}
\description{
Mean Estimated Data
}
