% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{fitmodel}
\alias{fitmodel}
\title{Fits Model using Stan}
\usage{
fitmodel(dpobj, stancode)
}
\arguments{
\item{dpobj}{'dpobj' from dataprocess()}

\item{stancode}{Stan code used to fit model.}
}
\value{
Model Fit
}
\description{
Fits Model using Stan
}
