% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gowers_distance}
\alias{gowers_distance}
\title{Calculate Gower's distance}
\usage{
gowers_distance(data)
}
\arguments{
\item{data}{Data.frame for distance calculation}
}
\value{
dist object containing distances
}
\description{
Calculate Gower's distance
}
\examples{
data <- data.frame(
V1 = c(1, 2, 3),
V2 = c(4, 5, 6),
V3 = c(7, 8, 9))
gowers_distance(data)
}
