% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{modelmTest}
\alias{modelmTest}
\title{Create ModelMatrix for Testing Data set}
\usage{
modelmTest(sumobj, cluster, time)
}
\arguments{
\item{sumobj}{'sumobj' from clustersum()}

\item{cluster}{cluster which the model is over}

\item{time}{time point which the model is over}
}
\value{
Model Matrix and True Testing Values
}
\description{
Create ModelMatrix for Testing Data set
}
