% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{parfit}
\alias{parfit}
\title{Fit Stan Models for each time point of Cluster in Parr}
\usage{
parfit(sumobj, stancode, cluster, ncores)
}
\arguments{
\item{sumobj}{'sumobj' from clustersum()}

\item{stancode}{.stan file with stancode for our fit}

\item{cluster}{which cluster to fit}

\item{ncores}{number of cores for stan fit}
}
\value{
list of models for cluster
}
\description{
Fit Stan Models for each time point of Cluster in Parr
}
