% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plots the dendrogram from HAC}
\usage{
plot_clusters(dendrogram, num_clus, pal)
}
\arguments{
\item{dendrogram}{A 'dendrogram' object as produced by as.dendrogram}

\item{num_clus}{The number of clusters to be shown}

\item{pal}{Colour palette used, usually a vector of string hex codes}
}
\value{
A plot of the dendrogram with the specified number of clusters
}
\description{
Plots the dendrogram from HAC
}
\examples{
hc <- hclust(dist(iris[, 1:4]))
dend <- as.dendrogram(hc)
pal <- c("#8DA0CB", "#E78AC3", "#66C2A5")
plot_clusters(dend, 3, pal)
}
