% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{plotpred}
\alias{plotpred}
\title{Plot Predict}
\usage{
plotpred(estobj, day)
}
\arguments{
\item{estobj}{'estobj' from estimate().}

\item{day}{day to estimate.}
}
\value{
ggplot of estimate and true values, RMSE, and data frame with estimate for day and true value (not centered). Class 'plotobj'
}
\description{
Plot Predict
}
