% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{weekly_profile}
\alias{weekly_profile}
\title{Calculate the weekly profile of a household}
\usage{
weekly_profile(data)
}
\arguments{
\item{data}{A list of data.frames in the same format as those in Irish.RData. Does not need to contain survey data.}
}
\value{
weekly_profile A tibble containing the weekly demand profiles of each household
}
\description{
Calculate the weekly profile of a household
}
\examples{
valid_input <- list(
extra = data.frame(tod = c(1, 2, 1, 2), dow = c("Mon", "Tue", "Mon", "Mon")),
indCons = data.frame(I1002 = c(1, 10, 20, 30), I1003 = c(40, 50, 60, 70)))
weekly_profile(valid_input)
}
