% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRCdemand-package.R
\docType{package}
\name{DRCdemand-package}
\alias{DRCdemand}
\alias{DRCdemand-package}
\title{DRCdemand: Forecasting of Irish Electricity Data}
\description{
Provides functions to aid analysis of the Irish electricity demand data set, including functions for clustering and forecasting.
}
\author{
\strong{Maintainer}: Codie Wood \email{codie.wood@bristol.ac.uk}

Authors:
\itemize{
  \item Dylan Dijk \email{zl22291@bristol.ac.uk}
  \item Rahil Morjaria \email{kl18909@bristol.ac.uk}
}

}
\keyword{internal}
