% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{totaldem}
\alias{totaldem}
\title{Generate DF of total demand, estimates and CI for Clustering Type for certain day}
\usage{
totaldem(estobj, no.clust, day)
}
\arguments{
\item{estobj}{object from estimate()}

\item{no.clust}{number of clusters in estobj}

\item{day}{day to get total demand, estimates and CI for.}
}
\value{
DF, and RMSE
}
\description{
Generate DF of total demand, estimates and CI for Clustering Type for certain day
}
